% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctbt_online.R
\name{ctbt_online}
\alias{ctbt_online}
\title{Continuous-time online backtest}
\usage{
ctbt_online(stock, w0 = 1000, p = 0.5, lambda = 0.5, timescale = 1/252)
}
\arguments{
\item{stock}{the stock prices}

\item{w0}{initial wealth to invest}

\item{p}{fraction determining training set}

\item{lambda}{EMA filter smoothing parameter}

\item{timescale}{timescale to scale the estimates by}
}
\description{
{An online backtest that fits models initially and then
updates it as it trades on out-of-sample data.
}
}
