% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtbt_offline.R
\name{dtbt_offline}
\alias{dtbt_offline}
\title{Discrete-time offline backtest}
\usage{
dtbt_offline(stock, w0 = 1000, p = 0.5, lambda = 0.5)
}
\arguments{
\item{stock}{the stock prices}

\item{w0}{initial wealth to invest}

\item{p}{fraction determining training set}

\item{lambda}{EMA filter smoothing parameter}
}
\description{
{An offline backtest that fits models once and then
backtests it on out-of-sample data.
}
}
