% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_backtest.R
\name{dtbt_online}
\alias{dtbt_online}
\title{Discrete-time online backtest}
\usage{
dtbt_online(stock, w0 = 1000, p = 0.5, lambda = 0.5)
}
\arguments{
\item{stock}{data-set of stock prices}

\item{w0}{initial wealth to invest}

\item{p}{fraction determining training set}

\item{lambda}{EMA filter smoothing parameter}
}
\description{
{An online backtest that fits models initially and then
updates it as it trades on out-of-sample data.
}
}
