% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema_dat_mat}
\alias{ema_dat_mat}
\title{EMA filter for DAT matrix of E((1+R_j)/(1+R_i))}
\usage{
ema_dat_mat(R, lambda = 0.94)
}
\arguments{
\item{R}{the data-set of arithmetic returns, multivariate time-series}

\item{lambda}{the smoothing parameter for the EMA filter}
}
\value{
matrix
}
\description{
{Estimate the mixed DAT matrix using an EMA filter
}
}
