% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ema.R
\name{ema_regression}
\alias{ema_regression}
\title{EMA filter for alphas and betas of a set of stocks (arithmetic returns)}
\usage{
ema_regression(R, lambda = 0.95)
}
\arguments{
\item{R}{the data-set of arithmetic returns, multivariate time-series, assumes the
first column is a market index like SPY etc.}

\item{lambda}{the smoothing parameter for the EMA filter}
}
\value{
vector
}
\description{
{Estimate the betas of a set of stocks using an EMA filter
}
}
