% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocators.R
\name{equal_weight}
\alias{equal_weight}
\title{The equal-weighted strategy}
\usage{
equal_weight(X)
}
\arguments{
\item{X}{data-set of arithmetic returns or log-returns}
}
\value{
vector
}
\description{
{The equally weighted diversified strategy.}
}
