% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocators.R
\name{kelly_gbm}
\alias{kelly_gbm}
\title{The Kelly-criterion strategy assuming a GBM model on log-returns}
\usage{
kelly_gbm(X, lambda = 0, h = 1/252)
}
\arguments{
\item{X}{data-set of periodic log-returns}

\item{lambda}{smoothing parameter for optional EMA filter}

\item{h}{timescale to use}
}
\value{
vector
}
\description{
{Maximize log growth by solving a QP problem.}
}
