% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocators.R
\name{kelly_index_reg}
\alias{kelly_index_reg}
\title{Optimal log growth under linear regression dynamics to market index}
\usage{
kelly_index_reg(R, lambda = 0, restraint = 1)
}
\arguments{
\item{R}{first column is assumed to be a market index, like SPY, QQQ, etc}

\item{lambda}{the smoothing parameter for the EMA filter}

\item{restraint}{portfolio restraint to pick}
}
\value{
list
}
\description{
{This uses the Taylor approximation and substitutes the linear
regression equations against a market index for the vector of returns.
}
}
