% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocators.R
\name{kelly_taylor}
\alias{kelly_taylor}
\title{Approximate log-optimal portfolio in discrete time with beta-hedging.}
\usage{
kelly_taylor(R, lambda = 0, betas = NULL, restraint = 1)
}
\arguments{
\item{R}{data-set of periodic arithmetic returns}

\item{lambda}{optional smoothing parameter for EMA, use 0 for naive estimates (default)}

\item{betas}{regression coefficients from a linear regression against market-index, or NULL}

\item{restraint}{the percentage of wealth to restrain to for investing}
}
\value{
list of allocations, growth-rate and volatility
}
\description{
{The same Taylor-approximation to the log-growth problem but with
beta neutral hedging as a constraint, as an option.}
}
