% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocators.R
\name{min_var}
\alias{min_var}
\title{Minimum variance portfolio}
\usage{
min_var(R, lambda = 0, restraint = 1)
}
\arguments{
\item{R}{data-set of periodic arithmetic returns}

\item{lambda}{optional smoothing parameter for EMA, use 0 for naive estimates (default)}

\item{restraint}{the percentage of wealth to restrain to for investing}
}
\value{
list of allocations, growth-rate and volatility
}
\description{
{Minimum variance portfolio.}
}
