% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{mv_solver}
\alias{mv_solver}
\title{QP solver for min 0.5 x^T A x-mu^t x subject to Gx >= b}
\usage{
mv_solver(drift, Sigma, betas = NULL, restraint = 1)
}
\arguments{
\item{drift}{Vector of means}

\item{Sigma}{Covariance matrix}

\item{betas}{optional betas}

\item{restraint}{restraint for portfolio investment}
}
\value{
list
}
\description{
{The Kelly-criterion in continuous time. A quadratic programming
problem with optional constraints for maximizing log-growth assuming a GBM model.
}
}
